Refactor Responses Download button to component (#2849)
* Refactor Download button and handler to component

* unused imports

* simplify signature

* remove comment

* unused import

* Redundant

* Move some checks up
